Installed and configured Pundit for authentication in different URLs. Corrected rubocop warnings. Corrected Wish rspec test that was failing from before.

+++ app/controllers/api/v1/rents_controller.rb
@@ -2,19 +2,18 @@ module Api
   module V1
     class RentsController < ApplicationController
       def index
-        render json: current_user.rents.page(params[:page])
+        render json: policy_scope(Rent).page(params[:page])
       end
 
       def show
-        @rent = rent
-        authorize @rent
+        authorize rent
         render json: rent
       end
 
       def create
         @rent = Rent.new(rent_params)
-        authorize @rent
-        if @rent.save
+        authorize rent
+        if rent.save
           head :created
         else
           render json: { error: rent.errors }, status: :unprocessable_entity
@@ -22,6 +21,7 @@ module Api
       end
 
       def destroy
+        authorize rent
         rent.destroy
         head :ok
       end
@@ -29,12 +29,13 @@ module Api
       private
 
       def rent_params
-        params.require(:rent).permit(:from, :to, :book_id)
+        params.require(:rent).permit(:from, :to, :book_id, :user_id)
       end
 
       def rent
         @rent ||= current_user.rents.find(params[:id])
       end
+
     end
   end
 end
